requirements.txt
# Multi-Tenant School Management System Requirements

# Flask and extensions
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography>=41.0.0

# Security
Werkzeug==3.0.1
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
python-dateutil>=2.8.0
requests>=2.31.0

# Development tools (optional)
Flask-DebugToolbar==0.13.1

# Production server (optional)
gunicorn==21.2.0

# Logging and monitoring
python-json-logger==2.0.7

# PDF Generation and Fee Management
reportlab==4.4.5
num2words==0.5.14
